% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkRank}
\alias{checkRank}
\title{Check rank}
\usage{
checkRank(taxid)
}
\arguments{
\item{taxid}{A character string with a taxid.}
}
\value{
A character string. NA if error.
}
\description{
\code{checkRank} checks that the rank matches the NCBI ID (taxids).
If not, NA is returned.
}
