% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkTaxonName}
\alias{checkTaxonName}
\title{Check taxon name}
\usage{
checkTaxonName(taxids)
}
\arguments{
\item{taxid}{A character string with a taxid.}
}
\value{
A character string. NA if error.
}
\description{
\code{checkTaxonName} checks that the taxon_name matches the NCBI ID (taxids).
If not, NA is returned.
}
