% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzzExports.R
\name{.getHeader}
\alias{.getHeader}
\title{Header for bugphyzz files}
\usage{
.getHeader(identifier = format(Sys.time(), "\%Y-\%m-\%d"))
}
\arguments{
\item{identifier}{any character vector. Defaults to today.}
}
\description{
Header for bugphyzz files
}
\examples{
bugphyzzExports:::.getHeader("3.17")
}
