% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzzExports.R
\name{.getSpecialThresholds}
\alias{.getSpecialThresholds}
\title{Returns a physiology's manually curated thresholds}
\usage{
.getSpecialThresholds(physiology)
}
\arguments{
\item{physiology}{the name of a physiology}
}
\value{
list of lists representing threshold ranges
}
\description{
Returns a physiology's manually curated thresholds
}
\examples{
.getSpecialThreshold()
}
