% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzzExports.R
\name{.hasSpecialThreshold}
\alias{.hasSpecialThreshold}
\title{Returns if a physiology has manually curated thresholds}
\usage{
.hasSpecialThreshold(physiology)
}
\arguments{
\item{physiology}{the name of a physiology}
}
\value{
boolean
}
\description{
Returns if a physiology has manually curated thresholds
}
\examples{
names(THRESHOLDS())
bugphyzzExports:::.hasSpecialThreshold(physiology = "coding genes")
bugphyzzExports:::.hasSpecialThreshold(physiology = "aerophilicity")
}
