% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzzExports.R
\name{.hasSpecialThresholds}
\alias{.hasSpecialThresholds}
\title{Returns if a physiology has manually curated thresholds}
\usage{
.hasSpecialThresholds(physiology)
}
\arguments{
\item{physiology}{the name of a physiology}
}
\value{
boolean
}
\description{
Returns if a physiology has manually curated thresholds
}
\examples{
.hasSpecialThreshold()
}
