% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzzExports.R
\name{.writeFileWithHeader}
\alias{.writeFileWithHeader}
\title{Write file given a signature with header}
\usage{
.writeFileWithHeader(signatures, file_path, header = header)
}
\arguments{
\item{signatures}{to write to file}

\item{file_path}{to write a file}

\item{header}{a character vector representing the header}
}
\description{
Write file given a signature with header
}
\examples{
writeFileWithHeader(signatures, "mysigs.gmt", .getHeader())
}
