% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzzExports.R
\name{.writeHeader}
\alias{.writeHeader}
\title{Write a header for a file given its path}
\usage{
.writeHeader(file_path, header = header)
}
\arguments{
\item{file_path}{path to the file}

\item{header}{a character vector representing the header}
}
\description{
Write a header for a file given its path
}
\examples{
# bugphyzzExports:::.writeHeader(file.path(tempdir(), "test.txt"))
}
