% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{homogenizeAerophilicityAttributeNames}
\alias{homogenizeAerophilicityAttributeNames}
\title{Homogenize attribute names of aerophilicity}
\usage{
homogenizeAerophilicityAttributeNames(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A data.frame.
}
\description{
\code{homogenizeAerophilicityAttribbuteNames} makes all levels in the
aerophilicity dataset at the same level in the GO tree.
}
