% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzzExports.R
\name{rangeToLogicalThr}
\alias{rangeToLogicalThr}
\title{Convert a range data.frame to logical based on thresholds}
\usage{
rangeToLogicalThr(df, thresholds)
}
\arguments{
\item{df}{A data.frame.}

\item{thresholds}{A list of thresholds.}
}
\value{
A data.frame.
}
\description{
\code{rangeToLogicalThr}
}
