% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeAccessionAndGenomeID}
\alias{removeAccessionAndGenomeID}
\title{Remove Accession_ID and Genome_ID columns}
\usage{
removeAccessionAndGenomeID(df)
}
\arguments{
\item{df}{A data.frame imported from bugphyzz.}
}
\value{
A data.frame.
}
\description{
\code{removeAcccessionAndGenomeID} removes Accession_ID and Genome_ID from
a bugphyzz dataset. The reason is that right now these columns can be
incomplete or inconsistent in some datasets, or just missing in some others.
I think a solution would be to implement a relational database in which we
have a data object (data.frame?) with all of the taxids.
}
