% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzzExports.R
\name{subsetByThreshold}
\alias{subsetByThreshold}
\title{Subset a data.frame by threshold}
\usage{
subsetByThreshold(df, thr)
}
\arguments{
\item{df}{A data.frame.}

\item{thr}{A named list of thresholds.}
}
\value{
A data.frame
}
\description{
Subset a data.frame by threshold
}
