% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzzExports.R
\name{THRESHOLDS}
\alias{THRESHOLDS}
\title{THRESHOLDS}
\usage{
THRESHOLDS()
}
\value{
A named nested list.
}
\description{
\code{THRESHOLDS} this function returns a list of thresholds for numeric
attributes. This woud allow to convert them to categorical/logical in
order to use them in the propagation workflow.
}
